Changed unnecessary static vars to member fields.
Instead of making variables in the PlayState static in order to access
them from the outside, you should use `(FlxG.state as
PlayState).field_or_method_name_here`. Those fields aren't *really*
static, and we shouldn't abuse the keyword.

Also, if you call `super` to call the overriden method, you should
always do it on the first line of the method you call it on.