Save player's progress in a shared object.
We can use the shared object so that the player can start again where
they left off. Note that it will start at the highest level that you
have played so go down to get to other levels. (We could disable the
loading part for now if you guys want.)